Cleaning up window and jquery
The vanilla version of iFrame Resizer inserts itself at window-level.
This is fine if you're using iFrame Resizer on your own site, but not
okay if you're using it on a different site.

By removing these lines, we can

Also removing a few tests because they're ancient and fail on travis.